Taken care of the 0.5 step on heater seekbar

+++ app/src/main/java/com/leprincesylvain/altentest/techinaltestv2/model/device/DeviceTableModel.kt
@@ -24,7 +24,7 @@ data class DeviceTableModel(
     var position: Int?,
 
     @ColumnInfo(name = "temperature")
-    var temperature: Int?,
+    var temperature: Double?,
 
     @ColumnInfo(name = "mode")
     var mode: String?
